+++ tapo-py/src/api.rs
@@ -7,7 +7,9 @@ mod light_handler;
 mod plug_energy_monitoring_handler;
 mod plug_handler;
 mod power_strip_handler;
+mod py_handler_ext;
 mod rgb_light_strip_handler;
+mod rgbic_light_strip_handler;
 
 pub use api_client::*;
 pub use child_devices::*;
@@ -18,4 +20,6 @@ pub use light_handler::*;
 pub use plug_energy_monitoring_handler::*;
 pub use plug_handler::*;
 pub use power_strip_handler::*;
+pub use py_handler_ext::*;
 pub use rgb_light_strip_handler::*;
+pub use rgbic_light_strip_handler::*;