+++ tapo-py/src/api/rgbic_light_strip_handler.rs
@@ -1,71 +1,72 @@
 use std::ops::{Deref, DerefMut};
 use std::sync::Arc;
 
+use pyo3::exceptions::PyTypeError;
 use pyo3::prelude::*;
 use pyo3::types::PyDict;
-use tapo::requests::Color;
-use tapo::responses::{DeviceInfoColorLightResult, DeviceUsageEnergyMonitoringResult};
-use tapo::{ColorLightHandler, HandlerExt};
+use tapo::requests::{Color, LightingEffect, LightingEffectPreset};
+use tapo::responses::{DeviceInfoRgbicLightStripResult, DeviceUsageEnergyMonitoringResult};
+use tapo::{HandlerExt, RgbicLightStripHandler};
 use tokio::sync::RwLock;
 
 use crate::api::PyHandlerExt;
 use crate::call_handler_method;
-use crate::requests::PyColorLightSetDeviceInfoParams;
+use crate::requests::{PyColorLightSetDeviceInfoParams, PyLightingEffect};
 
 #[derive(Clone)]
-#[pyclass(name = "ColorLightHandler")]
-pub struct PyColorLightHandler {
-    inner: Arc<RwLock<ColorLightHandler>>,
+#[pyclass(name = "RgbicLightStripHandler")]
+pub struct PyRgbicLightStripHandler {
+    pub inner: Arc<RwLock<RgbicLightStripHandler>>,
 }
 
-impl PyColorLightHandler {
-    pub fn new(handler: ColorLightHandler) -> Self {
+impl PyRgbicLightStripHandler {
+    pub fn new(handler: RgbicLightStripHandler) -> Self {
         Self {
             inner: Arc::new(RwLock::new(handler)),
         }
     }
 }
 
-impl PyHandlerExt for PyColorLightHandler {
+impl PyHandlerExt for PyRgbicLightStripHandler {
     fn get_inner_handler(&self) -> Arc<RwLock<(impl HandlerExt + 'static)>> {
         Arc::clone(&self.inner)
     }
 }
 
 #[pymethods]
-impl PyColorLightHandler {
+impl PyRgbicLightStripHandler {
     pub async fn refresh_session(&self) -> PyResult<()> {
         let handler = self.inner.clone();
         call_handler_method!(
             handler.write().await.deref_mut(),
-            ColorLightHandler::refresh_session,
+            RgbicLightStripHandler::refresh_session,
             discard_result
         )
     }
 
     pub async fn on(&self) -> PyResult<()> {
         let handler = self.inner.clone();
-        call_handler_method!(handler.read().await.deref(), ColorLightHandler::on)
+        call_handler_method!(handler.read().await.deref(), RgbicLightStripHandler::on)
     }
 
     pub async fn off(&self) -> PyResult<()> {
         let handler = self.inner.clone();
-        call_handler_method!(handler.read().await.deref(), ColorLightHandler::off)
+        call_handler_method!(handler.read().await.deref(), RgbicLightStripHandler::off)
     }
 
     pub async fn device_reset(&self) -> PyResult<()> {
         let handler = self.inner.clone();
         call_handler_method!(
             handler.read().await.deref(),
-            ColorLightHandler::device_reset
+            RgbicLightStripHandler::device_reset
         )
     }
 
-    pub async fn get_device_info(&self) -> PyResult<DeviceInfoColorLightResult> {
+    pub async fn get_device_info(&self) -> PyResult<DeviceInfoRgbicLightStripResult> {
         let handler = self.inner.clone();
         call_handler_method!(
             handler.read().await.deref(),
-            ColorLightHandler::get_device_info
+            RgbicLightStripHandler::get_device_info
         )
     }
 
@@ -73,7 +74,7 @@ impl PyColorLightHandler {
         let handler = self.inner.clone();
         let result = call_handler_method!(
             handler.read().await.deref(),
-            ColorLightHandler::get_device_info_json,
+            RgbicLightStripHandler::get_device_info_json,
         )?;
         Python::with_gil(|py| tapo::python::serde_object_to_py_dict(py, &result))
     }
@@ -82,7 +83,7 @@ impl PyColorLightHandler {
         let handler = self.inner.clone();
         call_handler_method!(
             handler.read().await.deref(),
-            ColorLightHandler::get_device_usage
+            RgbicLightStripHandler::get_device_usage
         )
     }
 
@@ -94,7 +95,7 @@ impl PyColorLightHandler {
         let handler = self.inner.clone();
         call_handler_method!(
             handler.read().await.deref(),
-            ColorLightHandler::set_brightness,
+            RgbicLightStripHandler::set_brightness,
             brightness
         )
     }
@@ -103,7 +104,7 @@ impl PyColorLightHandler {
         let handler = self.inner.clone();
         call_handler_method!(
             handler.read().await.deref(),
-            ColorLightHandler::set_color,
+            RgbicLightStripHandler::set_color,
             color
         )
     }
@@ -112,7 +113,7 @@ impl PyColorLightHandler {
         let handler = self.inner.clone();
         call_handler_method!(
             handler.read().await.deref(),
-            ColorLightHandler::set_hue_saturation,
+            RgbicLightStripHandler::set_hue_saturation,
             hue,
             saturation
         )
@@ -122,8 +123,36 @@ impl PyColorLightHandler {
         let handler = self.inner.clone();
         call_handler_method!(
             handler.read().await.deref(),
-            ColorLightHandler::set_color_temperature,
+            RgbicLightStripHandler::set_color_temperature,
             color_temperature
         )
     }
+
+    pub async fn set_lighting_effect(&self, lighting_effect: Py<PyAny>) -> PyResult<()> {
+        let handler = self.inner.clone();
+        let lighting_effect = map_lighting_effect(lighting_effect)?;
+        call_handler_method!(
+            handler.read().await.deref(),
+            RgbicLightStripHandler::set_lighting_effect,
+            lighting_effect
+        )
+    }
+}
+
+fn map_lighting_effect(lighting_effect: Py<PyAny>) -> PyResult<LightingEffect> {
+    if let Some(lighting_effect) =
+        Python::with_gil(|py| lighting_effect.extract::<LightingEffectPreset>(py).ok())
+    {
+        return Ok(lighting_effect.into());
+    }
+
+    if let Some(lighting_effect) =
+        Python::with_gil(|py| lighting_effect.extract::<PyLightingEffect>(py).ok())
+    {
+        return Ok(lighting_effect.into());
+    }
+
+    Err(PyErr::new::<PyTypeError, _>(
+        "Invalid lighting effect type. Must be one of `LightingEffect` or `LightingEffectPreset`",
+    ))
 }