+++ tapo-py/src/requests/set_device_info.rs
@@ -1,3 +1,5 @@
 mod color_light;
+mod lighting_effect;
 
 pub use color_light::*;
+pub use lighting_effect::*;