+++ tapo-py/src/requests/set_device_info/color_light.rs
@@ -1,9 +1,12 @@
 use std::ops::Deref;
 
+use pyo3::exceptions::PyTypeError;
 use pyo3::prelude::*;
 use tapo::requests::{Color, ColorLightSetDeviceInfoParams};
 
-use crate::api::{PyColorLightHandler, PyRgbLightStripHandler};
+use crate::api::{
+    PyColorLightHandler, PyHandlerExt, PyRgbLightStripHandler, PyRgbicLightStripHandler,
+};
 use crate::errors::ErrorWrapper;
 use crate::runtime::tokio;
 
@@ -20,32 +23,7 @@ impl PyColorLightSetDeviceInfoParams {
         }
     }
 
-    async fn _send_to_py_color_light_handler(&self, handler: PyColorLightHandler) -> PyResult<()> {
-        let params = self.params.clone();
-        let handler = handler.get_inner_handler();
-
-        tokio()
-            .spawn(async move {
-                let handler_lock = handler.read().await;
-
-                params
-                    .send(handler_lock.deref())
-                    .await
-                    .map_err(ErrorWrapper)?;
-
-                Ok::<_, ErrorWrapper>(())
-            })
-            .await
-            .map_err(anyhow::Error::from)
-            .map_err(ErrorWrapper::from)??;
-
-        Ok(())
-    }
-
-    async fn _send_to_py_rgb_light_strip_handler(
-        &self,
-        handler: PyRgbLightStripHandler,
-    ) -> PyResult<()> {
+    async fn _send_to_inner_handler(&self, handler: impl PyHandlerExt) -> PyResult<()> {
         let params = self.params.clone();
         let handler = handler.get_inner_handler();
 
@@ -110,17 +88,23 @@ impl PyColorLightSetDeviceInfoParams {
         if let Some(handler) =
             Python::with_gil(|py| handler.extract::<PyColorLightHandler>(py).ok())
         {
-            return self._send_to_py_color_light_handler(handler).await;
+            return self._send_to_inner_handler(handler).await;
         }
 
         if let Some(handler) =
             Python::with_gil(|py| handler.extract::<PyRgbLightStripHandler>(py).ok())
         {
-            return self._send_to_py_rgb_light_strip_handler(handler).await;
+            return self._send_to_inner_handler(handler).await;
         }
 
-        Err(PyErr::new::<pyo3::exceptions::PyTypeError, _>(
-            "Invalid handler type. Must be one of `PyColorLightHandler` or `PyRgbLightStripHandler`",
+        if let Some(handler) =
+            Python::with_gil(|py| handler.extract::<PyRgbicLightStripHandler>(py).ok())
+        {
+            return self._send_to_inner_handler(handler).await;
+        }
+
+        Err(PyErr::new::<PyTypeError, _>(
+            "Invalid handler type. Must be one of `PyColorLightHandler`, `PyRgbLightStripHandler` or `PyRgbicLightStripHandler`",
         ))
     }
 }