+++ tapo/src/requests.rs
@@ -1,12 +1,10 @@
 //! Tapo request objects.
 
-mod color;
 mod control_child;
 mod energy_data_interval;
 mod get_energy_data;
 mod get_trigger_logs;
 mod handshake;
-mod lighting_effect;
 mod login_device;
 mod multiple_request;
 mod secure_passthrough;
@@ -14,9 +12,7 @@ mod set_device_info;
 mod tapo_request;
 
 pub use crate::responses::TemperatureUnitKE100;
-pub use color::*;
 pub use energy_data_interval::*;
-pub use lighting_effect::*;
 pub use set_device_info::*;
 
 pub(crate) use control_child::*;