+++ tapo/src/requests/set_device_info.rs
@@ -1,9 +1,13 @@
+mod color;
 mod color_light;
 mod generic_device;
 mod light;
+mod lighting_effect;
 mod trv;
 
+pub use color::*;
 pub use color_light::*;
+pub use lighting_effect::*;
 
 pub(crate) use generic_device::*;
 pub(crate) use light::*;