+++ tapo/src/responses/device_info_result/rgbic_light_strip.rs
@@ -6,6 +6,7 @@ use crate::responses::{decode_value, DecodableResultExt, DefaultStateType, TapoR
 
 /// Device info of Tapo L920 and L930. Superset of [`crate::responses::DeviceInfoGenericResult`].
 #[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
 #[allow(missing_docs)]
 pub struct DeviceInfoRgbicLightStripResult {
     //
@@ -47,6 +48,18 @@ pub struct DeviceInfoRgbicLightStripResult {
     pub saturation: Option<u16>,
 }
 
+#[cfg(feature = "python")]
+#[pyo3::pymethods]
+impl DeviceInfoRgbicLightStripResult {
+    /// Gets all the properties of this result as a dictionary.
+    pub fn to_dict(&self, py: pyo3::Python) -> pyo3::PyResult<pyo3::Py<pyo3::types::PyDict>> {
+        let value = serde_json::to_value(self)
+            .map_err(|e| pyo3::exceptions::PyException::new_err(e.to_string()))?;
+
+        crate::python::serde_object_to_py_dict(py, &value)
+    }
+}
+
 impl TapoResponseExt for DeviceInfoRgbicLightStripResult {}
 
 impl DecodableResultExt for DeviceInfoRgbicLightStripResult {
@@ -58,16 +71,18 @@ impl DecodableResultExt for DeviceInfoRgbicLightStripResult {
     }
 }
 
-/// Color Light Strip Default State.
+/// RGB IC Light Strip Default State.
 #[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
 #[allow(missing_docs)]
 pub struct DefaultRgbicLightStripState {
     pub r#type: DefaultStateType,
     pub state: RgbicLightStripState,
 }
 
-/// Color Light Strip State.
+/// RGB IC Light Strip State.
 #[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
 #[allow(missing_docs)]
 pub struct RgbicLightStripState {
     pub brightness: Option<u8>,